requirements.txt
# Core Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# AI and LLM
openai==1.3.7
google-generativeai==0.3.2
langchain==0.0.350
langchain-openai==0.0.2
langchain-google-genai==0.0.5

# Vector Database and Embeddings
faiss-cpu==1.7.4
chromadb==0.4.18
sentence-transformers==2.2.2

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3

# Image Processing and OCR
opencv-python==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10
ultralytics==8.0.196

# Speech Processing
openai-whisper==20231117
SpeechRecognition==3.10.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
